L19

+++ app/Http/Controllers/LinkController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Link;
 use App\Rules\MyURL;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class LinkController extends Controller
@@ -15,6 +16,13 @@ class LinkController extends Controller
     public function index()
     {
         $links = Link::orderBy('clicks', 'desc')->get();
+
+        return view('welcome', compact('links'));
+    }
+
+    public function dashboard()
+    {
+        $links = Auth::user()->links()->get();
 
         return view('links.index', compact('links'));
     }
@@ -59,6 +67,7 @@ class LinkController extends Controller
         $link = Link::create([
             'full_url' => $request->link,
             'shortened_url' => $shortened,
+            'user_id' => Auth::user()?->id,
         ]);
 
 
@@ -92,9 +101,15 @@ class LinkController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(string $slug)
     {
-        //
+        $link = Link::slug($slug)->first();
+
+        if ($link->user_id != Auth::user()->id) {
+            return back();
+        }
+
+        return view('links.edit', compact('link'));
     }
 
     /**
@@ -102,7 +117,26 @@ class LinkController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
+        $request->validate([
+            'id' => 'required|integer',
+            'full_url' => [
+                'required',
+                new MyURL,
+            ],
+        ]);
+
+        $link = Link::findOrFail($request->id);
+
+        if ($link->user_id != Auth::user()->id) {
+            return back();
+        }
+
+        $link->update([
+            'full_url' => $request->full_url,
+            'clicks' => 0,
+        ]);
+
+        return redirect('/');
     }
 
     /**
@@ -115,6 +149,11 @@ class LinkController extends Controller
         ]);
 
         $link = Link::findOrFail($request->id);
+
+        if ($link->user_id != Auth::user()->id) {
+            return back();
+        }
+
         $link->delete();
 
         return redirect('/');
